refactor!: move types in a single declaration file
