+++ src/tests/bot/setup.spec.ts
@@ -1,13 +1,7 @@
+import setup from '@/preference/setup.ts'
 import { MemorySessionStorage, R } from '@deps'
-import {
-  afterAll,
-  afterEach,
-  assertEquals,
-  beforeEach,
-  describe,
-  it,
-} from '@dev_deps'
-import { SessionData } from '@/types/sessionData.ts'
+import { afterAll, assertEquals, beforeEach, describe, it } from '@dev_deps'
+import { SessionData } from '@/types.ts'
 import {
   chat,
   from,
@@ -30,8 +24,6 @@ beforeEach(() => {
 afterAll(async () => {
   await new Promise((r) => setTimeout(r, 0))
 })
-
-import setup from '@/preference/setup.ts'
 
 describe('setup', () => {
   it('saves preferences', async () => {
