--- src/types/sessionData.ts
@@ -1,12 +1,0 @@
-import { Context, Conversation, ConversationFlavor, SessionFlavor } from '@deps'
-
-export interface SessionData {
-  preferences: string[]
-}
-
-export type PreferencesContext =
-  & Context
-  & SessionFlavor<SessionData>
-  & ConversationFlavor
-
-export type ConversationContext = Conversation<PreferencesContext>
